Less dependence on global 'monitors' variable.
There is a global variable that holds an array of monitors at the time the
extension was initialized. This variable becomes out of sync with the rest of
gnome.

This change does not eliminate the monitors variable variable, but it reduces
its use. When a user enters a key binding associated with a preset grid
location, the set of actual monitors is used.