Let user cycle through list of preset grid sizes.
Old form for preset: "2x2 0:0 1:0"
New form for preset: "2x2 0:0 0:1, 3x3 0:0 0:2, 3x3 0:0 1:2"

The function that applies the preset now checks the window position against the
rectangles specified in the preset array. If the rectangle matches one of them,
the next preset is used. This allows the user to cycle through sizes with a
single accelerator binding.

The implementation uses a new TypeScript file (tilespec.ts) to parse a preset
and apply it to the working area of the display.