Add keyboard shortcut for moving window to next monitor.
Gnome offers a shortcut for moving a monitor to the monitor to the left, right,
up, or down. This requires 4 keybindings. This approach requires only one
keybinding. Also, in my experience, it is less mental effort to repeatedly hit
the "move-to-next-monitor" shortcut repeatedly.

The change adds a new tilespec.ts file for parsing gTile-formatted rectangle
specifications. This was copied from the "bazelify" branch of
github.com/gonzojive/gTile. Future refactoring should use this for parsing and
printing tile specification strings instead of the current code in app.ts.